feat(BinaryDataset): add load_from_trainset
Given a existing trainset, build a binary dataset from it. Note that the
item's columns and rows will follow the internal trainset representation
and will almost certainly differ from the original dataset from which
the trainset was derived. Use the trainset's to_raw_* methods to convert
correctly between this spaces.